fix: enforce DNS-only container authorities
