+++ test/container-entrypoint.test.mjs
@@ -1,6 +1,9 @@
 import assert from 'node:assert/strict';
 import { spawnSync } from 'node:child_process';
 import { EventEmitter } from 'node:events';
+import { mkdtempSync, rmSync, symlinkSync } from 'node:fs';
+import { tmpdir } from 'node:os';
+import { join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import test from 'node:test';
 
@@ -8,17 +11,21 @@ import {
   installPublicAuthorityMarker,
   isHelpRequest,
   validatedEngineArgs,
-} from '../docker/entrypoint.mjs';
+} from '../docker/entrypoint-core.mjs';
 
 const entrypoint = fileURLToPath(
   new URL('../docker/entrypoint.mjs', import.meta.url),
 );
 const acceptedArgs = ['--base-url', 'https://relay.ratifyprotocol.com'];
 
-function run(args) {
-  return spawnSync(process.execPath, [entrypoint, ...args], {
+function runAt(executable, args) {
+  return spawnSync(process.execPath, [executable, ...args], {
     encoding: 'utf8',
   });
+}
+
+function run(args) {
+  return runAt(entrypoint, args);
 }
 
 function assertAccepted(args = acceptedArgs) {
@@ -52,12 +59,56 @@ test('refuses a single-label --base-url hostname', () => {
 });
 
 test('refuses a loopback --base-url', () => {
+  assertRefused(
+    ['--base-url', 'https://relay.localhost'],
+    /must not use localhost/,
+  );
   assertRefused(['--base-url', 'https://127.0.0.1'], /loopback IP address/);
   assertRefused(
     ['--base-url', 'https://[::ffff:127.0.0.1]'],
     /loopback IP address/,
   );
   assertAccepted(); // Control: a non-loopback hostname succeeds.
+});
+
+test('refuses the special-use .local namespace', () => {
+  assertRefused(
+    ['--base-url', 'https://relay.local'],
+    /special-use \.local namespace/,
+  );
+  assertAccepted(); // Control: a public DNS authority succeeds.
+});
+
+test('refuses IPv4 and IPv6 literal authorities', () => {
+  assertRefused(
+    ['--base-url', 'https://203.0.113.5'],
+    /DNS name; IP literals are forbidden/,
+  );
+  assertRefused(
+    ['--base-url', 'https://[2001:db8::1]'],
+    /DNS name; IP literals are forbidden/,
+  );
+  assertAccepted(); // Control: a DNS authority succeeds.
+});
+
+test('refuses an invalid authority when invoked through a symlink', (t) => {
+  const directory = mkdtempSync(join(tmpdir(), 'relaycast-entrypoint-'));
+  const symlink = join(directory, 'entrypoint.mjs');
+  symlinkSync(entrypoint, symlink);
+  t.after(() => rmSync(directory, { force: true, recursive: true }));
+
+  const refused = runAt(symlink, [
+    '--base-url',
+    'http://relay.ratifyprotocol.com',
+  ]);
+  assert.equal(refused.status, 64, refused.stderr);
+  assert.match(refused.stderr, /refused to start/);
+  assert.match(refused.stderr, /must use https/);
+
+  // Control: the same symlink runs the executable's successful help path.
+  const accepted = runAt(symlink, ['--help']);
+  assert.equal(accepted.status, 0, accepted.stderr);
+  assert.match(accepted.stdout, /Required public HTTPS origin/);
 });
 
 test("accepts Ratify's HTTPS deployment authority", () => {
